sql: clean up scan and merge join ordering code
Remove the code that was used by the HP to initialize orderings for
scans and merge joins.

Release note: None
