+++ pkg/sql/join.go
@@ -35,7 +35,6 @@ type joinNode struct {
 	// on the equality columns (or a subset of them). The column indices refer to
 	// equality columns: a ColIdx of i refers to left column
 	// pred.leftEqualityIndices[i] and right column pred.rightEqualityIndices[i].
-	// See computeMergeJoinOrdering. This information is used by distsql planning.
 	mergeJoinOrdering sqlbase.ColumnOrdering
 
 	props physicalProps
