+++ pkg/sql/physical_props.go
@@ -689,185 +689,6 @@ func (pp *physicalProps) applyExpr(evalCtx *tree.EvalContext, expr tree.TypedExp
 	pp.notNullCols.UnionWith(extractNotNullConstraints(expr))
 }
 
-// computeMergeJoinOrdering determines if merge-join can be used to perform a join.
-//
-// It takes the orderings of the two data sources that are to be joined on a set
-// of equality columns (the join condition is that the value for the column
-// colA[i] equals the value for column colB[i]).
-//
-// If merge-join can be used, the function returns a ColumnOrdering that refers
-// to the equality columns by their index in colA/ColB. Specifically column i in
-// the returned ordering refers to column colA[i] for A and colB[i] for B. This
-// is the ordering that must be used by the merge-join.
-//
-// The returned ordering can be partial, i.e. only contains a subset of the
-// equality columns. This indicates that a hybrid merge/hash join can be used
-// (or alternatively, an extra sorting step to complete the ordering followed by
-// a merge-join). See example below.
-//
-// Note that this function is not intended to calculate the output physicalProps
-// of joins (this is a separate problem with other complications).
-//
-// Examples:
-//  -  natural join between
-//       table A with columns (u, v, x, y)  primary key x+,y+,u+
-//       table B with columns (x, y, w)     primary key x+,y+
-//     equality columns are x, y
-//     a ordering is 2+,3+,0+
-//     b ordering is 0+,1+
-//     colA is {2, 3}   // column indices of x,y in table A
-//     colB is {0, 1}   // column indices of x,y in table B
-//
-//     The function returns 0+,1+. This result maps to ordering 2+,3+ for A and
-//     0+,1+ for B; this is what the merge-join will use: it will interleave
-//     rows by comparing column A2 with column B0, breaking equalities by
-//     comparing column A3 with column B1.
-//
-//  -  natural join between
-//       table A with columns (u, v, x, y)  primary key x+
-//       table B with columns (x, y, w)     primary key x+,y+
-//     equality columns are x, y
-//     a ordering is 2+
-//     b ordering is 0+,1+
-//     colA is {2, 3}   // column indices of x,y in table A
-//     colB is {0, 1}   // column indices of x,y in table B
-//
-//     The function returns 0+. This maps to ordering 2+ for A and 0+ for B.
-//     This is a partial ordering, so a hybrid merge-join can be used: groups of
-//     rows that are equal on columns a2 and b0 are loaded and a hash-join is
-//     performed on this group. Alternatively, an extra sorting step could be
-//     used to refine the ordering (this sorting step would also use the partial
-//     ordering to only sort within groups) followed by a regular merge-join.
-func computeMergeJoinOrdering(a, b physicalProps, colA, colB []int) sqlbase.ColumnOrdering {
-	if len(colA) != len(colB) {
-		panic(fmt.Sprintf("invalid column lists %v; %v", colA, colB))
-	}
-	if a.isEmpty() || b.isEmpty() || len(colA) == 0 {
-		return nil
-	}
-
-	var result sqlbase.ColumnOrdering
-
-	// First, find any merged columns that are constant in both sources. This
-	// means that in each source, this column only sees one value.
-	for i := range colA {
-		if a.constantCols.Contains(colA[i]) && b.constantCols.Contains(colB[i]) {
-			// The direction here is arbitrary - the orderings guarantee that either works.
-			// TODO(radu): perhaps the correct thing would be to return an
-			// physicalProps with this as a constant column.
-			result = append(result, sqlbase.ColumnOrderInfo{ColIdx: i, Direction: encoding.Ascending})
-		}
-	}
-
-	// To understand what's going on, it's useful to first think of the easy
-	// case: there are no constant columns or equivalent column groups, we just
-	// have simple orderings. In this case we need to check that:
-	//  - the first column in A's ordering is an equality column, and
-	//  - the first column in B's ordering is the same equality column.
-	//    If this is the case, we can check the same for the second column, and so
-	//    on. If not, we stop.
-	//
-	// This gets more complicated because of constant columns. If the first
-	// column in A's ordering is an equality column and the corresponding B column
-	// is a constant, this pairing also works. This means that we will not
-	// necessarily consume the orderings at the same rate. The code below
-	// tracks the remaining parts of the orderings in ordA/ordB.
-	//
-	// Another complication is the "key" flag: such an ordering remains correct
-	// when appending arbitrary columns to it.
-	//
-	// Column groups complicate things further, for example:
-	//   A: 1=3; 2=4; 1+,2+
-	//   B: 1+, 2+, 3+, 4+
-	// Here we match (1/3)+ with 1+, then (2/4)+ with 2+, after which column 3+
-	// matches a column inside an earlier group and is thus inconsequential for
-	// A's ordering (like a constant column would be). Note that the direction of
-	// a redundant column can even differ, B: 1+, 2+, 3-, 4- would match with A
-	// just as well. This is because within each group of rows with the same
-	// values on columns 1 and 2 in A, there is a single value for columns 3 and
-	// 4.
-	//
-	// To help handle these cases in a unified manner, we keep a list of "seen"
-	// column groups on each side that can be used arbitrarily to extend an
-	// ordering: these are the constant column groups and the groups processed so
-	// far.
-
-	seenGroupsA := a.constantCols.Copy()
-	seenGroupsB := b.constantCols.Copy()
-
-MainLoop:
-	for ordA, ordB := a.ordering, b.ordering; ; {
-		doneA, doneB := (len(ordA) == 0), (len(ordB) == 0)
-		// See if the first column group in each ordering contain the same equality
-		// column.
-		if !doneA && !doneB {
-			foundCol := -1
-			groupA := a.eqGroups.Find(ordA[0].ColIdx)
-			groupB := b.eqGroups.Find(ordB[0].ColIdx)
-			for i := range colA {
-				if a.eqGroups.Find(colA[i]) == groupA && b.eqGroups.Find(colB[i]) == groupB {
-					// Both ordering groups contain the i-th equality column.
-					foundCol = i
-					break
-				}
-			}
-			if foundCol != -1 {
-				dir := ordA[0].Direction
-				if dir != ordB[0].Direction {
-					// Both orderings start with the same merged column, but the
-					// ordering is different. That's all, folks.
-					break MainLoop
-				}
-				result = append(result, sqlbase.ColumnOrderInfo{ColIdx: foundCol, Direction: dir})
-				seenGroupsA.Add(groupA)
-				seenGroupsB.Add(groupB)
-				ordA, ordB = ordA[1:], ordB[1:]
-				continue MainLoop
-			}
-		}
-		// See if any column in the first group in A is "seen" in B. Or, if
-		// we consumed B and it is a "key", then we are free to add any other
-		// columns in A.
-		//
-		// For example, assuming a "natural join" of columns 1 through 4:
-		//   A: 1+,2+,3+,4+
-		//   B: 1=3; 1+,2+,4+
-		// After we match 1+,2+ in A with 1+,2+ in B, column 3 in A is already
-		// "seen" in B (same group with 1).
-		if !doneA {
-			groupA := a.eqGroups.Find(ordA[0].ColIdx)
-			for i := range colA {
-				if a.eqGroups.Find(colA[i]) == groupA &&
-					((doneB && b.isKey(seenGroupsB)) ||
-						seenGroupsB.Contains(b.eqGroups.Find(colB[i]))) {
-					result = append(result, sqlbase.ColumnOrderInfo{ColIdx: i, Direction: ordA[0].Direction})
-					seenGroupsA.Add(groupA)
-					ordA = ordA[1:]
-					continue MainLoop
-				}
-			}
-		}
-		// See if any column in the first group in B is "seen" in A. Or, if
-		// we consumed A and it is a "key", then we are free to add any other
-		// columns in B. This case is symmetric to the case above.
-		if !doneB {
-			groupB := b.eqGroups.Find(ordB[0].ColIdx)
-			for i := range colB {
-				if b.eqGroups.Find(colB[i]) == groupB &&
-					((doneA && a.isKey(seenGroupsA)) ||
-						seenGroupsA.Contains(a.eqGroups.Find(colA[i]))) {
-					result = append(result, sqlbase.ColumnOrderInfo{ColIdx: i, Direction: ordB[0].Direction})
-					seenGroupsB.Add(groupB)
-					ordB = ordB[1:]
-					continue MainLoop
-				}
-			}
-		}
-		break
-	}
-	return result
-}
-
 // getColVarIdx detects whether an expression is a straightforward
 // reference to a column or index variable. In this case it returns
 // the index of that column's in the descriptor's []Column array.
