+++ pkg/sql/physical_props_test.go
@@ -13,7 +13,6 @@ package sql
 import (
 	"fmt"
 	"math/rand"
-	"reflect"
 	"testing"
 
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
@@ -556,202 +555,6 @@ func TestTrimOrdering(t *testing.T) {
 			tc.props.trim(tc.desired)
 			if !propsEqual(tc.props, tc.expected) {
 				t.Errorf("expected %s, got %s", tc.expected.AsString(nil), tc.props.AsString(nil))
-			}
-		})
-	}
-}
-
-func TestComputeMergeJoinOrdering(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	testCases := []struct {
-		name       string
-		a, b       physicalProps
-		colA, colB []int
-		expected   sqlbase.ColumnOrdering
-	}{
-		{
-			name: "basic",
-			a: makePhysProps(
-				makeColumnOrdering(1, asc, 2, desc, 3, asc),
-			),
-			b: makePhysProps(
-				makeColumnOrdering(3, asc, 4, desc),
-			),
-			colA:     []int{1, 2},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(0, asc, 1, desc),
-		},
-
-		{
-			name: "groups",
-			a: makePhysProps(
-				equivGroups{{2, 5}},
-				makeColumnOrdering(1, asc, 2, desc, 3, asc),
-			),
-			b: makePhysProps(
-				equivGroups{{2, 3}},
-				makeColumnOrdering(2, asc, 4, desc),
-			),
-			colA:     []int{1, 5},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(0, asc, 1, desc),
-		},
-
-		{
-			name: "const-a",
-			a: makePhysProps(
-				constCols{1, 2},
-			),
-			b: makePhysProps(
-				makeColumnOrdering(3, asc, 4, desc),
-			),
-			colA:     []int{1, 2},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(0, asc, 1, desc),
-		},
-
-		{
-			name: "const-b",
-			a: makePhysProps(
-				makeColumnOrdering(1, asc, 2, desc, 3, asc),
-			),
-			b: makePhysProps(
-				constCols{3, 4},
-			),
-			colA:     []int{1, 2},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(0, asc, 1, desc),
-		},
-
-		{
-			name: "const-with-groups",
-			a: makePhysProps(
-				equivGroups{{1, 4, 5}},
-				makeColumnOrdering(1, asc, 2, desc, 3, asc),
-			),
-			b: makePhysProps(
-				constCols{3, 4},
-			),
-			colA:     []int{5, 2},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(0, asc, 1, desc),
-		},
-		{
-			name: "const-both",
-			a: makePhysProps(
-				constCols{1},
-				makeColumnOrdering(2, desc),
-			),
-			b: makePhysProps(
-				constCols{4},
-				makeColumnOrdering(3, asc),
-			),
-			colA:     []int{1, 2},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(1, desc, 0, asc),
-		},
-		{
-			name: "key-a",
-			a: makePhysProps(
-				makeColumnOrdering(1, asc),
-				weakKeys{{1}},
-				notNullCols{1},
-			),
-			b: makePhysProps(
-				makeColumnOrdering(3, asc, 4, desc),
-			),
-			colA:     []int{1, 2},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(0, asc, 1, desc),
-		},
-
-		{
-			name: "key-b",
-			a: makePhysProps(
-				makeColumnOrdering(1, asc, 2, desc),
-			),
-			b: makePhysProps(
-				makeColumnOrdering(3, asc),
-				weakKeys{{3}},
-				notNullCols{3},
-			),
-			colA:     []int{1, 2},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(0, asc, 1, desc),
-		},
-
-		{
-			name: "partial-ordering-1",
-			a: makePhysProps(
-				makeColumnOrdering(1, asc, 3, asc, 2, desc),
-			),
-			b: makePhysProps(
-				makeColumnOrdering(3, asc, 4, desc),
-			),
-			colA:     []int{1, 2},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(0, asc),
-		},
-		{
-			name: "partial-ordering-2",
-			a: makePhysProps(
-				makeColumnOrdering(1, asc, 2, desc, 3, asc),
-			),
-			b: makePhysProps(
-				makeColumnOrdering(3, asc, 5, desc, 4, desc),
-			),
-			colA:     []int{1, 2},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(0, asc),
-		},
-		{
-			name: "partial-ordering-with-groups",
-			a: makePhysProps(
-				equivGroups{{1, 8}, {3, 9}},
-				makeColumnOrdering(1, asc, 3, asc, 2, desc),
-			),
-			b: makePhysProps(
-				equivGroups{{3, 5}},
-				makeColumnOrdering(3, asc, 4, desc),
-			),
-			colA:     []int{8, 2},
-			colB:     []int{3, 4},
-			expected: makeColumnOrdering(0, asc),
-		},
-		{
-			name: "grouped-cols",
-			a: makePhysProps(
-				equivGroups{{0, 2}, {1, 3}},
-				makeColumnOrdering(2, asc, 3, asc),
-			),
-			b: makePhysProps(
-				makeColumnOrdering(0, asc, 1, asc, 2, asc, 3, asc),
-			),
-			colA:     []int{0, 1, 2, 3},
-			colB:     []int{0, 1, 2, 3},
-			expected: makeColumnOrdering(0, asc, 1, asc, 2, asc, 3, asc),
-		},
-		{
-			name: "grouped-cols-opposite-order",
-			a: makePhysProps(
-				equivGroups{{0, 2}, {1, 3}},
-				makeColumnOrdering(2, asc, 1, asc),
-			),
-			b: makePhysProps(
-				makeColumnOrdering(0, asc, 1, asc, 2, desc, 3, desc),
-			),
-			colA:     []int{0, 1, 2, 3},
-			colB:     []int{0, 1, 2, 3},
-			expected: makeColumnOrdering(0, asc, 1, asc, 2, desc, 3, desc),
-		},
-	}
-	for i := range testCases {
-		tc := testCases[i]
-		t.Run(tc.name, func(t *testing.T) {
-			result := computeMergeJoinOrdering(tc.a, tc.b, tc.colA, tc.colB)
-			if !reflect.DeepEqual(tc.expected, result) {
-				t.Errorf("expected %v, got %v", tc.expected, result)
 			}
 		})
 	}
